+++ Kaguya/Kaguya/Graphics/AssetManager.cpp
@@ -79,7 +79,7 @@ void AssetManager::Initialize()
 								static_cast<UINT64>(Metadata.width),
 								static_cast<UINT>(Metadata.height),
 								static_cast<UINT16>(Metadata.arraySize),
-								Metadata.mipLevels);
+								static_cast<UINT16>(Metadata.mipLevels));
 							break;
 
 						case TEX_DIMENSION::TEX_DIMENSION_TEXTURE3D:
